box model test case modified to test guid schema
This is the box model from the

Here is the email from Bob Lipman:
Hi everyone,

I created a version of SFA that works with the updated schema.  I also created a sample STEP file with one instance each of the five new entity types.  Attached is the STEP file and resulting SFA spreadsheet.

Please let me know if there are other example STEP files that use the new entities.

Thanks,

Bob

Signed-off-by: TRThurman <thomas.r.thurman@imonmail.com>