added examples from Mikael Hedlind ISO_14405-1_2016_Figure_17a
These examples were originally created as part of ISO 10303 work on bug 7505 but interesting effects occur when a simple diameter change occurs.
There is an internal groove with diameter 60 mm.
The "-0" is the original design with outer diameter of 72 mm
The "-1" is one design iteration with outer diameter of 64 mm
The "-2" is a second design iteration with outer diameter of 56 mm that splits the solid into two bodies.

Signed-off-by: TRThurman <thomas.r.thurman@imonmail.com>